Upgrade libmodbus version 3.0.6 to 3.1.4

+++ include/coriental_motor.h
@@ -44,7 +44,7 @@ private:
   bool m_valid;
   
 public:
-  COrientalMotor(std::string, uint16_t, char, uint8_t, uint8_t, uint8_t id);
+  COrientalMotor(std::string, unsigned int, char, unsigned short, unsigned short, unsigned short id);
   //int set_goal_velocity()
 
   int move_absolute_position();
